requirements.txt
numpy
scikit-learn
faiss-cpu
sentence-transformers
torch
transformers
PyPDF2
accelerate
requests
python-dotenv
